# openapi/glacier-notes-v1.yaml
openapi: 3.1.0
info:
  title: Glacier Notes Cloud API
  version: 1.0.0
  description: |
    Canonical contract for Glacier Notes Cloud. Operations are added before their
    implementing milestone. UUIDs use RFC 4122 form and timestamps are ISO-8601 UTC.
servers:
  - url: /
...
paths:
...
  /api/v1/me/profile:
    get:
      tags: [Current User]
      summary: Get the current user's profile and self-service capabilities
      operationId: getCurrentUserProfile
      responses:
        '200':
          description: Current profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
        default: { $ref: '#/components/responses/ProblemResponse' }
    patch:
      tags: [Current User]
      summary: Update the current user's username or display name
      operationId: updateCurrentUserProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserProfileUpdate' }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
        '409': { $ref: '#/components/responses/ProblemResponse' }
        '422': { $ref: '#/components/responses/ProblemResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
  /api/v1/me/password:
    put:
      tags: [Current User]
      summary: Change the current user's password and revoke all sessions
      operationId: changeCurrentUserPassword
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordChangeRequest' }
      responses:
        '204': { description: Password changed and all sessions revoked }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '422': { $ref: '#/components/responses/ProblemResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
  /api/v1/me/email-change:
    post:
      tags: [Current User]
      summary: Request verification of a new email address
      operationId: requestCurrentUserEmailChange
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailChangeRequest' }
      responses:
        '202': { description: Verification email accepted for delivery }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '409': { $ref: '#/components/responses/ProblemResponse' }
        '422': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/RateLimitedProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
  /api/v1/me/settings:
    get:
      tags: [Current User]
      summary: Get synchronized current-user preferences
      operationId: getCurrentUserSettings
      responses:
        '200':
          description: Current preferences and policy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSettings' }
        default: { $ref: '#/components/responses/ProblemResponse' }
    patch:
      tags: [Current User]
      summary: Update synchronized current-user preferences
      operationId: updateCurrentUserSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserSettingsUpdate' }
      responses:
        '200':
          description: Updated preferences and policy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSettings' }
        '422': { $ref: '#/components/responses/ProblemResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
  /api/v1/me/deletion:
    post:
      tags: [Current User]
      summary: Irreversibly request deletion of the current account
      operationId: deleteCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SelfDeletionRequest' }
      responses:
        '202': { description: Irreversible deletion accepted }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '409': { $ref: '#/components/responses/ProblemResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
...
  /api/v1/admin/users/{userId}/deletion:
    post:
      tags: [Administration]
      summary: Schedule retained or immediate user deletion
      operationId: scheduleUserDeletion
      parameters: [{ $ref: '#/components/parameters/UserId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminDeletionRequest' }
      responses:
        '202':
          description: Deletion scheduled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUser' }
        '409': { $ref: '#/components/responses/ProblemResponse' }
        '422': { $ref: '#/components/responses/ProblemResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
...
  /api/v1/images:
    post:
      tags: [Images]
      summary: Upload and normalize an owned image
      operationId: uploadImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Normalized image metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImageMetadata' }
        '413': { $ref: '#/components/responses/ProblemResponse' }
        '422': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
  /api/v1/images/{imageId}:
    parameters:
      - name: imageId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Images]
      summary: Stream an owned normalized image
      operationId: downloadImage
      responses:
        '200':
          description: Normalized image bytes
          content:
            image/png: { schema: { type: string, format: binary } }
            image/jpeg: { schema: { type: string, format: binary } }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
    delete:
      tags: [Images]
      summary: Delete an unreferenced owned image immediately
      operationId: deleteImage
      responses:
        '204': { description: Image deleted }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        '409': { $ref: '#/components/responses/ProblemResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
  /api/v1/images/{imageId}/thumbnail:
    get:
      tags: [Images]
      summary: Stream an owned image thumbnail
      operationId: downloadImageThumbnail
      parameters:
        - name: imageId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Thumbnail bytes
          content:
            image/png: { schema: { type: string, format: binary } }
            image/jpeg: { schema: { type: string, format: binary } }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
  /api/v1/images/{imageId}/metadata:
    get:
      tags: [Images]
      summary: Get owned image metadata
      operationId: getImageMetadata
      parameters:
        - name: imageId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Image metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImageMetadata' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
  /api/v1/me/storage:
    get:
      tags: [Current User]
      summary: Get current user's image-storage usage
      operationId: getCurrentUserStorage
      responses:
        '200':
          description: Current storage usage and quota
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageUsage' }
        default: { $ref: '#/components/responses/ProblemResponse' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: GLACIER_SESSION
      description: Server-managed session cookie.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on state-changing cookie-authenticated requests.
    bootstrapToken:
      type: apiKey
      in: header
      name: X-Bootstrap-Token
      description: One-time deployment secret used only while the instance is uninitialized.
  parameters:
...
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  headers:
    CorrelationId:
      description: Request correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9._-]+$'
  responses:
    ProblemResponse:
      description: Standardized API error
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    RateLimitedProblemResponse:
      description: Too many attempts
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
        Retry-After:
          description: Seconds until another attempt is permitted.
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    UserProfile:
      type: object
      required: [id, username, email, role, emailChangeAvailable, selfDeletionEnabled]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        email: { type: string, format: email }
        displayName: { type: string }
        role: { type: string, enum: [USER, ADMIN] }
        emailChangeAvailable: { type: boolean }
        selfDeletionEnabled: { type: boolean }
    UserProfileUpdate:
      type: object
      properties:
        username: { type: string, minLength: 3, maxLength: 64, pattern: '^[\p{L}\p{N}._-]+$' }
        displayName: { type: string, maxLength: 128 }
    PasswordChangeRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, format: password, minLength: 1, maxLength: 128 }
        newPassword: { type: string, format: password, minLength: 12, maxLength: 128, pattern: '^\S+$' }
    EmailChangeRequest:
      type: object
      required: [currentPassword, newEmail]
      properties:
        currentPassword: { type: string, format: password, minLength: 1, maxLength: 128 }
        newEmail: { type: string, format: email, maxLength: 320 }
    UserSettings:
      type: object
      required: [theme, language, moveCheckedToBottom, trashAutoPurgeDays, trashAutoPurgeMayBeDisabled]
      properties:
        theme: { type: string, enum: [dark, light] }
        language: { type: string, enum: [en, de] }
        moveCheckedToBottom: { type: boolean }
        trashAutoPurgeDays: { type: integer, minimum: 0, maximum: 3650 }
        trashAutoPurgeMayBeDisabled: { type: boolean }
    UserSettingsUpdate:
      type: object
      properties:
        theme: { type: string, enum: [dark, light] }
        language: { type: string, enum: [en, de] }
        moveCheckedToBottom: { type: boolean }
        trashAutoPurgeDays: { type: integer, minimum: 0, maximum: 3650 }
    SelfDeletionRequest:
      type: object
      required: [currentPassword]
      properties:
        currentPassword: { type: string, format: password, minLength: 1, maxLength: 128 }
    AdminDeletionRequest:
      type: object
      required: [mode]
      properties:
        mode: { type: string, enum: [RETAINED, IMMEDIATE] }
        confirmation: { type: string, maxLength: 64 }
    AdminUser:
      type: object
      required: [id, username, email, role, status, createdAt, storageBytes, noteCount, notebookCount, imageCount]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        email: { type: string, format: email }
        displayName: { type: string }
        role: { type: string, enum: [USER, ADMIN] }
        status: { type: string, enum: [PENDING_ACTIVATION, ACTIVE, LOCKED, DEACTIVATED, PENDING_DELETION, DELETED] }
        createdAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time }
        pendingDeletionAt: { type: string, format: date-time }
        deletionDueAt: { type: string, format: date-time }
        storageBytes: { type: integer, format: int64, minimum: 0 }
        noteCount: { type: integer, format: int64, minimum: 0 }
        notebookCount: { type: integer, format: int64, minimum: 0 }
        imageCount: { type: integer, format: int64, minimum: 0 }
...
    ImageMetadata:
      type: object
      required: [id, mimeType, byteSize, width, height, thumbnailMimeType, thumbnailByteSize, thumbnailWidth, thumbnailHeight, contentUrl, thumbnailUrl, createdAt]
      properties:
        id: { type: string, format: uuid }
        mimeType: { type: string, enum: [image/png, image/jpeg] }
        byteSize: { type: integer, format: int64, minimum: 0 }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
        thumbnailMimeType: { type: string, enum: [image/png, image/jpeg] }
        thumbnailByteSize: { type: integer, format: int64, minimum: 0 }
        thumbnailWidth: { type: integer, minimum: 1 }
        thumbnailHeight: { type: integer, minimum: 1 }
        contentUrl: { type: string, format: uri-reference }
        thumbnailUrl: { type: string, format: uri-reference }
        createdAt: { type: string, format: date-time }
    StorageUsage:
      type: object
      required: [usedBytes, quotaBytes, remainingBytes, imageCount]
      properties:
        usedBytes: { type: integer, format: int64, minimum: 0 }
        quotaBytes: { type: integer, format: int64, minimum: 0 }
        remainingBytes: { type: integer, format: int64, minimum: 0 }
        imageCount: { type: integer, format: int64, minimum: 0 }
...
    ProblemDetails:
      type: object
      required: [type, title, status, detail, correlationId, errorCode]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        correlationId:
          type: string
        errorCode:
          type: string
          pattern: '^[A-Z][A-Z0-9_]+$'
        validationErrors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        currentVersion:
          type: integer
          format: int64
          minimum: 0
        currentUpdatedAt:
          type: string
          format: date-time
      example:
        type: https://glacier-notes.example/problems/content-version-conflict
        title: Content Changed
        status: 409
        detail: The content changed since it was read.
        instance: /api/v1/notes/80e10c3b-aaba-4f1c-a427-7564b1d5eaaf
        correlationId: 35907ba3-c92f-46bb-84f0-936325851b91
        errorCode: CONTENT_VERSION_CONFLICT
        validationErrors: []
        currentVersion: 4
    ValidationError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...
security:
  - sessionCookie: []
    csrfHeader: []